just using uint32 for masking buttons
git-svn-id: https://chan-sccp-b.svn.sourceforge.net/svnroot/chan-sccp-b/trunk@1982 fa19cda8-b258-4bce-aa1a-843e42da0b19

+++ src/sccp_mwi.c
@@ -473,8 +473,8 @@ void sccp_mwi_check(sccp_device_t *device)
 	sccp_line_t		*line = NULL;
 	sccp_moo_t 		*r;
 
-	uint8_t 			status;
-	uint64_t 		mask;
+	uint8_t			status;
+	uint32_t 		mask;
 
 	if(!device)
 		return;
